refactor(cli): extract diagnostics command module
Moves the src/core/settings static import out of the CLI entry's import
graph and into the only command group that uses it.
